+++ R/mrgsolve.R
@@ -93,8 +93,6 @@ validate_idata <- function(idata) {
 ##'
 ##' \itemize{
 ##' 
-##' \item \code{mtime} numeric vector of times where the model is evaluated 
-##' (with solver reset), but results are not included in simulated output
 ##' 
 ##' \item \code{Request} a vector of compartment or table names to take in 
 ##' simulated output; if this is specified, \code{request} is ignored
@@ -123,8 +121,15 @@ validate_idata <- function(idata) {
 ##' \item \code{tad} logical; when \code{TRUE} a column is added to simulated 
 ##' output is added showing the time since the last dose.  Only data records 
 ##' with \code{evid == 1} will be considered doses for the purposes of 
-##' \code{tad} calculation.
-##' 
+##' \code{tad} calculation. The \code{tad} can be properly calculated with 
+##' a dosing lag time in the model as long as the dosing lag time (specified 
+##' in \code{$MAIN})  is always  appropriate for any subsequent doses scheduled 
+##' through \code{addl}.  This will always be true if the lag time doesn't 
+##' change over time.  But it might (possibly) not hold if the lag time changes
+##' prior to the last dose in the \code{addl} sequence.  This known limitation
+##' shouldn't affect \code{tad} calculation in most common dosing lag time
+##' implementations.  
+##'  
 ##' \item \code{nocb} if \code{TRUE} (default), time-varying items in a data 
 ##' set will be implemented as next observation carried back; if \code{FALSE} 
 ##' time-varying items in a data set will be implemented as last observation 
@@ -397,7 +402,6 @@ mrgsim_nid <- function(x, nid, events = ev(), ...) {
 ##' @param data a simulation data set
 ##' @param idata individual-level data
 ##' @param carry.out data items to copy into the output
-##' @param mtime not used
 ##' @param seed deprecated
 ##' @param Request compartments or captured variables to retain
 ##' in the simulated output
@@ -426,7 +430,6 @@ do_mrgsim <- function(x,
                       data,
                       idata = null_idata,
                       carry.out = character(0),
-                      mtime = numeric(0),
                       seed = as.integer(NA),
                       Request = character(0),
                       capture = NULL,
@@ -536,7 +539,7 @@ do_mrgsim <- function(x,
   parin$recsort <- recsort
   parin$obsonly <- obsonly
   parin$obsaug <- obsaug
-  parin$mtime <- sort(unique(mtime))
+  #parin$mtime <- sort(unique(mtime))
   parin$filbak <- filbak
   parin$tad <- tad
   parin$nocb <- nocb
@@ -688,7 +691,7 @@ do_mrgsimple <- function(x,
   parin$recsort <- recsort
   parin$obsonly <- obsonly
   parin$obsaug <- obsaug
-  parin$mtime <- numeric(0)
+  #parin$mtime <- numeric(0)
   parin$filbak <- filbak
   parin$tad <- tad
   parin$nocb <- nocb